requirements.txt
certifi==2018.4.16
chardet==3.0.4
colorlog==3.1.4
Django==2.1.1
django-cors-headers==2.4.0
django-extensions==2.1.2
django-filter==2.0.0
django-oauth-toolkit==1.2.0
djangorestframework==3.8.2
idna==2.7
oauthlib==2.1.0
Pillow==5.2.0
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
django-otp==0.5.0
qrcode==6.0
rest-condition==1.0.3
psycopg2==2.7.5
psycopg2-binary==2.7.5
django-storages==1.7.1
google-cloud-storage==1.12.0
gevent==1.3.5
gunicorn==19.9.0